+++ app/_layout.tsx
@@ -1,10 +1,32 @@
 import { Tabs } from "expo-router";
-import { FunctionComponent } from "react";
+import { FC, FunctionComponent } from "react";
 import Icon from "react-native-vector-icons/MaterialCommunityIcons";
 import AppWrapper, { theme } from "../src/AppWrapper";
+import ActionsPanel from "../src/ActionsPanel";
+import { Divider, Text } from "react-native-paper";
+import { View } from "react-native";
+import { ItemStatus } from "../store/ItemList.types";
 interface LayoutProps {}
 
 const iconSize = 41;
+
+type HeaderProps = {
+  title: string;
+  type: ItemStatus;
+};
+
+const Header: FC<HeaderProps> = ({ title, type }) => (
+  <View style={{ margin: 16 }}>
+    <Text
+      variant="titleLarge"
+      style={{ textAlign: "center", fontWeight: "bold" }}
+    >
+      {title}
+    </Text>
+    <ActionsPanel type={type} />
+    <Divider bold />
+  </View>
+);
 
 const Layout: FunctionComponent<LayoutProps> = () => {
   return (
@@ -29,6 +51,7 @@ const Layout: FunctionComponent<LayoutProps> = () => {
         <Tabs.Screen
           name="shopping-list"
           options={{
+            header: () => <Header title="Shopping List" type="shoppingList" />,
             tabBarIcon: ({ color }) => {
               return (
                 <Icon name="clipboard-list" size={iconSize} color={color} />
@@ -39,6 +62,7 @@ const Layout: FunctionComponent<LayoutProps> = () => {
         <Tabs.Screen
           name="fridge"
           options={{
+            header: () => <Header title="Fridge" type="fridge" />,
             tabBarIcon: ({ color }) => {
               return <Icon name="fridge" size={iconSize} color={color} />;
             },
@@ -47,6 +71,7 @@ const Layout: FunctionComponent<LayoutProps> = () => {
         <Tabs.Screen
           name="food-diary"
           options={{
+            header: () => <Header title="Food Diary" type="foodDiary" />,
             tabBarIcon: ({ color }) => {
               return <Icon name="food" size={iconSize} color={color} />;
             },
