+++ app/fridge.tsx
@@ -2,14 +2,10 @@ import { FunctionComponent } from "react";
 import ItemList from "../src/ItemsList";
 import { ScrollView } from "react-native";
 import globalStyles from "../utils/globalStyles";
-import { Text } from "react-native-paper";
 
-interface FridgeProps {}
-
-const Fridge: FunctionComponent<FridgeProps> = () => {
+const Fridge: FunctionComponent = () => {
   return (
     <ScrollView contentContainerStyle={globalStyles.listContainer}>
-      <Text variant="titleMedium">Your fridge:</Text>
       <ItemList type="fridge" />
     </ScrollView>
   );
